Remove rendering from span node
A first step to improve rendering of the guides and separate it from
parsing was to extract the code from the node span node creation. It looks
like the rending wasn't required at all by the code. As far as we can
trust psalm and phpstan on this.

A second step was to clean up the constructor of the spannode as we want it to be.
The new create method should me moved to some kind of parser rule in the future,
but since it doesn't require renderer, it's not needed for me right now to improve this
even more.